backend/requirements.txt
Chameleon==2.11
Cython==0.19.1
Django==1.5.1
Fabric==1.5.3
Jinja2==2.6
Mako==0.8.0
MarkupSafe==0.15
PIL==1.1.7
PasteDeploy==1.5.0
PyRSS2Gen==1.0.0
Pygments==1.5
South==0.7.6
Sphinx==1.1.3
Twisted==12.0.0
URLObject==2.0.2
WebOb==1.2.3
Werkzeug==0.8.3
altgraph==0.9
amqp==1.0.11
amqplib==1.0.2
anyjson==0.3.3
bdist-mpkg==0.4.4
billiard==2.7.3.28
bonjour-py==0.3
celery==3.0.13
coverage==3.5.2
distribute==0.6.36
django-celery==3.0.9
django-cors-headers==0.06
django-debug-toolbar==0.9.4
django-extensions==0.9
django-jenkins==0.12.1
django-mptt==0.5.5
django-nose==1.1
django-tastypie==0.9.14
django-uni-form==0.9.0
djangorestframework==2.2.4
docutils==0.9.1
fuzzywuzzy==0.1
-e git://github.com/surfly/gevent.git@454a77ca561868854760b2d9cbfa3bf3bbd2e062#egg=gevent-dev
gevent-zeromq==0.2.5
graphviz==0.2.2
greenlet==0.4.1
gunicorn==0.14.6
hashids==0.8.3
httplib2==0.7.5
ipdb==0.6.1
ipython==0.12.1
isoweek==1.2.0
johnny-cache==1.4
kombu==2.5.10
lettuce==0.2.5
lettuce-webdriver==0.2
logilab-astng==0.23.1
logilab-common==0.57.1
macholib==1.4.2
mimeparse==0.1.3
mock==1.0.1
modulegraph==0.9.1
nose==1.2.1
numpy==1.6.1
paramiko==1.10.1
pep8==1.3.3
psycopg2==2.4.5
py2app==0.6.3
pyOpenSSL==0.13
pycrypto==2.6
pyflakes==0.5.0
pygame==1.9.2pre
pygraphviz==1.2
pylint==0.25.1
pyobjc-core==2.3.2a0
pyobjc-framework-AddressBook==2.3.2a0
pyobjc-framework-AppleScriptKit==2.3.2a0
pyobjc-framework-AppleScriptObjC==2.3.2a0
pyobjc-framework-Automator==2.3.2a0
pyobjc-framework-CFNetwork==2.3.2a0
pyobjc-framework-CalendarStore==2.3.2a0
pyobjc-framework-Cocoa==2.3.2a0
pyobjc-framework-Collaboration==2.3.2a0
pyobjc-framework-CoreData==2.3.2a0
pyobjc-framework-CoreLocation==2.3.2a0
pyobjc-framework-CoreText==2.3.2a0
pyobjc-framework-DictionaryServices==2.3.2a0
pyobjc-framework-ExceptionHandling==2.3.2a0
pyobjc-framework-FSEvents==2.3.2a0
pyobjc-framework-InputMethodKit==2.3.2a0
pyobjc-framework-InstallerPlugins==2.3.2a0
pyobjc-framework-InstantMessage==2.3.2a0
pyobjc-framework-InterfaceBuilderKit==2.3.2a0
pyobjc-framework-LatentSemanticMapping==2.3.2a0
pyobjc-framework-LaunchServices==2.3.2a0
pyobjc-framework-Message==2.3.2a0
pyobjc-framework-OpenDirectory==2.3.2a0
pyobjc-framework-PreferencePanes==2.3.2a0
pyobjc-framework-PubSub==2.3.2a0
pyobjc-framework-QTKit==2.3.2a0
pyobjc-framework-Quartz==2.3.2a0
pyobjc-framework-ScreenSaver==2.3.2a0
pyobjc-framework-ScriptingBridge==2.3.2a0
pyobjc-framework-SearchKit==2.3.2a0
pyobjc-framework-ServerNotification==2.3.2a0
pyobjc-framework-ServiceManagement==2.3.2a0
pyobjc-framework-SyncServices==2.3.2a0
pyobjc-framework-SystemConfiguration==2.3.2a0
pyobjc-framework-WebKit==2.3.2a0
pyobjc-framework-XgridFoundation==2.3.2a0
pyramid==1.4.1
pysqlite==2.6.3
python-dateutil==1.5
python-digest==1.7
python-memcached==1.48
pyzmq==2.2.0
reportlab==2.5
repoze.lru==0.6
selenium==2.24.0
simplejson==2.6.0
ssh==1.7.14
stevedore==0.8
sure==0.10.3
translationstring==1.1
uWSGI==1.2.3
urllib3==1.4
urwid==1.1.1
vboxapi==1.0
venusian==1.0a8
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.0
wsgiref==0.1.2
xattr==0.6.2
xlrd==0.9.2
zope.deprecation==4.0.2
zope.interface==4.0.5
